+++ test/unittest/source/domain/test_assassinate.cpp
@@ -78,6 +78,8 @@ domain:
 
       }
 
+
+
    } //test::domain::assassinate
 
 } // casual